+++ dataSources/ncbi/biosample/augment.py
@@ -1,6 +1,6 @@
 import pandas as pd
-import lib.dataframeFuncs as dff
-import lib.commonFuncs as cmn
+import lib.dataframes as dff
+import lib.common as cmn
 
 def augmentBiosample(df: pd.DataFrame) -> pd.DataFrame:
     return dff.splitField(df, "ncbi_lat long", cmn.latlongToDecimal, ["decimalLatitude", "decimalLongitude"])
