+++ src/lib/__init__.py
@@ -1,4 +1,4 @@
-import lib.tools.logger as lg
+import lib.logger as lg
 import lib.config as cfg
 
 lg.createLogger(logToConsole=cfg.Settings.logToConsole, logLevel=cfg.Settings.logLevel)
